requirements.txt
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
facebook-sdk==2.0.0
Faker==0.8.10
Flask==0.12.2
Flask-DebugToolbar==0.10.1
Flask-OAuth==0.12
Flask-SQLAlchemy==2.3.2
httplib2==0.10.3
idna==2.6
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauth2==1.9.0.post1
pkg-resources==0.0.0
psycopg2==2.7.3.2
pysqlite==2.8.3
python-dateutil==2.6.1
pytz==2017.3
pyzipcode==1.0
requests==2.9.1
six==1.11.0
SQLAlchemy==1.2.2
text-unidecode==1.1
urllib3==1.22
Werkzeug==0.14.1
